Move [warmup-]run logic into dedicated function
