+++ crates/app/src/app.rs
@@ -1,9 +1,10 @@
-use std::path::PathBuf;
+use std::{path::PathBuf, time::Instant};
 
 use graph::prelude::*;
 
 use clap::AppSettings::DeriveDisplayOrder;
 use kommandozeile::*;
+use log::info;
 
 mod loading;
 mod page_rank;
@@ -113,4 +114,35 @@ enum Algorithm {
         #[clap(long)]
         undirected: bool,
     },
+}
+
+pub(crate) fn time(runs: usize, warmup_runs: usize, f: impl Fn()) {
+    for run in 1..=warmup_runs {
+        let start = Instant::now();
+        f();
+        let took = start.elapsed();
+
+        info!(
+            "Warm-up run {} of {} finished in {:.6?}",
+            run, warmup_runs, took,
+        );
+    }
+
+    let mut durations = vec![];
+
+    for run in 1..=runs {
+        let start = Instant::now();
+        f();
+        let took = start.elapsed();
+        durations.push(took);
+
+        info!("Run {} of {} finished in {:.6?}", run, runs, took,);
+    }
+
+    let total = durations
+        .into_iter()
+        .reduce(|a, b| a + b)
+        .unwrap_or_default();
+
+    info!("Average runtime: {:?}", total / runs as u32);
 }
