+++ crates/app/src/triangle_count.rs
@@ -2,7 +2,6 @@ use graph::prelude::*;
 
 use log::info;
 use std::path::Path as StdPath;
-use std::time::Instant;
 
 use super::*;
 
@@ -64,34 +63,9 @@ where
         relabel_graph(&mut graph);
     }
 
-    for run in 1..=warmup_runs {
-        let start = Instant::now();
+    time(runs, warmup_runs, || {
         global_triangle_count(&graph);
-        let took = start.elapsed();
-
-        info!(
-            "Warm-up run {} of {} finished in {:.6?}",
-            run, warmup_runs, took,
-        );
-    }
-
-    let mut durations = vec![];
-
-    for run in 1..=runs {
-        let start = Instant::now();
-        global_triangle_count(&graph);
-        let took = start.elapsed();
-        durations.push(took);
-
-        info!("Run {} of {} finished in {:.6?}", run, runs, took);
-    }
-
-    let total = durations
-        .into_iter()
-        .reduce(|a, b| a + b)
-        .unwrap_or_default();
-
-    info!("Average runtime: {:?}", total / runs as u32);
+    });
 
     Ok(())
 }
